when fixing stuff, please fix it all!
git-svn-id: https://ilk.uvt.nl/svn/trunk/sources/Timbl6@14649 12f355fe-0486-481a-ad91-c297ab22b4e3

+++ src/Instance.cxx
@@ -1201,7 +1201,7 @@ namespace Timbl {
     // otherwise 1. Special case when reading the TopDistribution.
     //
     TargetValue *target;
-    int freq;
+    size_t freq;
     unsigned int index;
     double sw;
     char nextCh;
